Create catdoor_final
I needed to add a file to my GitHub so I could create my first GitHub Pages template. So, I've uploaded some code I wrote for a project in my ME106 Fundamentals of Mechatronics Engineering class this past spring. I'm not satisfied with the quality of this project, but this is simply the most recent code I have on hand so it serves as a benchmark for where I want to improve from.